refactored tot_pos_login_organism
